+++ CHANGES.md
@@ -25,6 +25,10 @@
 - [UPDATE] 前面カメラ・背面カメラの切り替えボタンアイコンを置き換える
   - `camera` から `camera.rotate` に置き換える
   - @t-miya
+- [ADD] 接続メニューに `接続時カメラ有効` 項目を追加する
+  - DataChannel、Simulcast、Spotlight、VideoChat が対象
+  - `無効` で接続した場合はカメラハードミュート状態で開始する
+  - @t-miya
 - [ADD] カメラハードミュート機能を追加する
   - DataChannel、Simulcast、Spotlight、VideoChat が対象
   - 録画 -> ソフトミュート -> ハードミュート -> 録画 と遷移するようにカメラミュートボタンの挙動を変更する